pixel-agnostic ResizeProcessor
